wasmbadge: more WIP

+++ wasmbadge/modules/image/image.go
@@ -5,19 +5,19 @@ import (
 )
 
 //go:wasmimport image new
-func new_image(ptr, sz uint32) uint32
+func new_image(ptr, sz int32) int32
 
 //go:wasmimport image show
-func image_show(p uint32) uint32
+func image_show(p int32) int32
 
 type Image struct {
-	Ref uint32
+	Ref int32
 }
 
 func NewImage(data []byte) *Image {
 	ptr, sz := convert.BytesToWasmPtr(data)
 	return &Image{
-		Ref: new_image(ptr, sz),
+		Ref: new_image(int32(ptr), int32(sz)),
 	}
 }
